+++ product-mini/platforms/linux/main.c
@@ -2,5 +2,5 @@
  * Copyright (C) 2019 Intel Corporation.  All rights reserved.
  * SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
  */
-
+#include "wali.h"
 #include "../posix/main.c"
